requirements.txt
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
SQLAlchemy==2.0.7
Jinja2==3.1.2
pytz==2022.7.1
PyRSS2Gen==1.1.0
Werkzeug==2.2.3
python-dotenv==1.0.0